Added key-based merging for arrays of maps:
Using the "(( merge ))" token for an array merge
will check to see if the source and destination elements
of the array are maps, containing the "name" key. If so,
it will merge elements together (regardless of order),
identifying matches based on their "name".

Using "(( merge on <mykey> ))" will behave the same way,
but use the key represented by <mykey> to determine object
matching.

Existing elements are kept in order, and updated in place (if
there was a matching object in the new array). New elements that
do not match any existing elements are appended to the end of the
array (in the order specified in the new array).