New process for embedding examples in README.md
The `./ci/embed-examples' Perl script looks through README.md and find
all references to the contents of `examples/**/*.yml`, replacing the
full YAML code block with the contents of the example file.

This helps to keep the README in sync with the examples in examples/,
which are in turn tested regeularly (to make sure they work).