Optimize (and warn for) Expr Short-Circuits
If an operator specifies an expression operand that short circuits (that
is, resolves to a literla while there are more alternates), spruce will
print a warning and remove the dead code from the expression.  This
helps keep the data-flow graph clean and relevant.

For example:

    ---
    meta:
      domain: example.com
      fallback:
        domain: (( grab meta.domain ))

    domain: (( grab meta.domain || "nx.domain" || meta.fallback.domain ))

will emit a warning that the literal value "nx.domain" short-circuits
the expression, and the `$.domain' key will *not* have a dependency on
`$.meta.fallback.domain', since the third alternate will never be
activated.