Handle VAULT_TOKEN better for interactive use
If VAULT_TOKEN is not found in the environment (or is empty), and there
exists a file $HOME/.vault-token, read that file and use its contents as
the token.  This file is created by a call to `vault auth`.

This still supports the CI pipeline use case, but makes things easier on
humans trying to generate sensitive BOSH manifests with Spruce + Vault

Fixes #90