Always set VAULT_TOKEN header, even during redirects
In HA-modes, targetting a non-leader Vault causes that API endpoint to
issue 307 Temporary Redirects directing the client at the correct
(leader) API endpoint.  The default redirection logic in net/http fails
to set the X-Vault-Token header on these subsequent requests, so we
override the policy function here and explicitly set it.

Fixes #91