Switch to Gox
Using gox for cross-compiling instead of goxc gets us:

  - No more tarballs (just binaries in the GH releases)
  - An easier integration with jumpbox scripts
  - Statically-compiled binaries (for Alpine Linux / musl!)

To that end, this commit introduces several changes to the build scripts
(both build.sh and shipit.sh) and the pipeline to support gox.  Some of
these changes are cleanups / clarifications as I found the shipit script
in particular to be difficult to follow.