Fix Cursor.Glob aliasing bug
Fixes #100

Cursors were being inadvertantly aliased when globbing Kleene-closures
that occurred as the last component at a list.

This caused the `(( static_ips ... ))` operator to fail horribly when
handling  multiple static IP ranges, because it would keep duplicating
the last path and incorrectly generate the pool of IP addresses.  In
effect, this:

    networks:
      - name: net1
        # ...
        static:
          - 10.0.0.5 - 10.0.0.6
          - 10.0.0.8 - 10.0.0.9
          - 10.0.0.100 - 10.0.0.199

    jobs:
      - name: api
        networks:
          - name: net1
            static_ips: (( static_ips 0 ))

would misconstrue the static ranges as:

  - 10.0.0.100 - 10.0.0.199
  - 10.0.0.100 - 10.0.0.199
  - 10.0.0.100 - 10.0.0.199

miscalculate the pool size as 3 * 100, and assign 10.0.0.100 to the job.

This has now been fixed.