Better Vault auth/unauth handling
`spruce merge` now requires an explicit environment variable ($REDACT)
to be set, or it will fail to render the final YAML if it contains any
`(( vault ... ))` calls.  This should help cases where people
accidentally forget to authenticate to the vault, and then deploy BOSH
manifests that are REDACTED all over the place.