Honor TLS/SSL verification settings from `safe -k`
re: https://github.com/starkandwayne/safe/releases/tag/v0.0.28

As of safe release 0.0.28, targeting a Vault with `-k` will persist for
future operations.  This is done by way of the `~/.svtoken` file.

This commit introduces support for that new addition, and causes Spruce
to behave according to the `safe` operator's wishes.