Key merge no longer panics if target keys' value is sequence or hash
If implicit merge got us here, a warning is given, and it falls back to an inline strategy.
If explicitly stated to merge, an error is given.
A new WarningError object with a regulated warning output was made to facilitate this. The operator warnings were updated to use this.
Warnings can be silenced from the code, so no more warnings from passing tests!
Tests were added.
Log library scope changes to facilitate testing with new warning system.