Document + test `((!...))` syntax for credhub passthrough
Spruce ignores `((!password))` because operators must start with
ASCII [a-zA-Z].  BOSH implemented a workaround whereby `((!...))` gets
interpreted as `((...))` to allow a bug-workaround in Spiff to make use
of Credhub.  This commit documents and tests the suitability of that
same workaround for Spruce.

(This functionality has existed, undocumented, since Spruce v0)

Fixes #216