Added a test case to the main test cases list, which merges two files from a assets directory to show that: If you add a `(( prune ))` in a list of maps structure that is later merged with another file, which itself changes the order in that list, you will end up deleting the wrong parts of the YAML structure: https://play.spruce.cf/#06832a74c691b5eaf823edfb702a9f41
In order to fix the prune in list of maps structure problem, we have to store the path not by using the index of the list entry, but rather the actual key name that identifies the list entry. However, this breaks another test case that checks for a specific string. That's why the other test case was modified in order to check for the path using the key name in it.