Fix issue 267: different merging behaviour
Remove full starting block of function `mergeArray`, which used to cover
operations `inline`, `replace`, and (default) `merge on key`. Their
individual actions are now part of the general array modification
routines.

Introduce new type called modificationType, which represents the type of
array operation that is possible, for example: insert, delete, merge
and so on. Add a set of constants for the respective operation types.

Remove `modificationDefinition` element `delete` in favour for the more
general modificationType, which also covers deletions now.

Update test cases that used `modificationDefinition` element `delete`
directly and now rely on the `modificationType`.