Add load operator to Spruce
Add code and test cases for `(( load ... ))` operator, which interprets the
provided literal argument as a file location. The file is either loaded
locally from the file system or as a URI remotely. The file content must be
valid YAML or JSON and is then inserted into the tree structure.