Adds spruce fan command for multi-doc yamling
`spruce fan` will take a source file and multiple target files (multi-doc YAML supported!),
and merge the target documents on top of the source, outputting a giant multi-doc YAML stream.
It takes all the same flags as a spruce merge, and supports merging the same way, with the
exception that only two documents are ever merged together (source, and each target doc independently).